reorganize views in App.js and rename projectcard to projectCard
