+++ frontend/src/components/projectCard/nCardPlaceholder.js
