+++ frontend/src/components/projectCard/projectProgressBar.js
