+++ frontend/src/views/welcome.js
@@ -9,8 +9,8 @@ import { HelpCard, FirstProjectBanner } from '../components/user/content';
 import { calculateCompleteness } from '../components/user/completeness';
 import { UserInformationForm } from '../components/user/forms';
 import { WelcomeCard } from '../components/user/content';
-import { ProjectCard } from '../components/projectcard/projectCard';
-import { nCardPlaceholders } from '../components/projectcard/nCardPlaceholder';
+import { ProjectCard } from '../components/projectCard/projectCard';
+import { nCardPlaceholders } from '../components/projectCard/nCardPlaceholder';
 import { useFetch } from '../hooks/UseFetch';
 import { useSetTitleTag } from '../hooks/UseMetaTags';
